refactor: change google auth route
