+++ src/routes/v1/index.ts
@@ -6,9 +6,11 @@ import orderRouter from './order';
 import billRouter from './bill';
 import paymentRouter from './payment';
 import authenticationMiddleware from '../../middlewares/authentication';
+import authRouter from './auth';
 
 const v1Router = express.Router()
 
+v1Router.use('/auth', authRouter);
 v1Router.use('/user', authenticationMiddleware, userRouter);
 v1Router.use('/friend', authenticationMiddleware, friendRouter);
 v1Router.use('/friends-order', authenticationMiddleware, friendsOrderRouter);
